Address ssh-tmux teardown review findings
