+++ Packages/macOS/CmuxControlSocket/Sources/CmuxControlSocket/Coordinator/Workspace/ControlWorkspaceStrings.swift
@@ -11,6 +11,8 @@ public struct ControlWorkspaceStrings: Sendable, Equatable {
     /// `workspace.closeProtected.message` — the `workspace.close` protected-pin
     /// error.
     public let closeProtected: String
+    /// The `workspace.close` local-teardown failure message.
+    public let closeFailed: String
     /// `socket.workspace.reorderMany.missingOrder`.
     public let reorderManyMissingOrder: String
     /// `socket.workspace.reorderMany.duplicateWorkspace`.
@@ -26,6 +28,7 @@ public struct ControlWorkspaceStrings: Sendable, Equatable {
     ///
     /// - Parameters:
     ///   - closeProtected: The `workspace.close` protected-pin message.
+    ///   - closeFailed: The `workspace.close` local-teardown failure message.
     ///   - reorderManyMissingOrder: The missing-order message.
     ///   - reorderManyDuplicateWorkspace: The duplicate-workspace message.
     ///   - reorderManyWorkspaceNotFound: The workspace-not-found message.
@@ -33,6 +36,7 @@ public struct ControlWorkspaceStrings: Sendable, Equatable {
     ///   - reorderManyTabManagerUnavailable: The TabManager-unavailable message.
     public init(
         closeProtected: String,
+        closeFailed: String,
         reorderManyMissingOrder: String,
         reorderManyDuplicateWorkspace: String,
         reorderManyWorkspaceNotFound: String,
@@ -40,6 +44,7 @@ public struct ControlWorkspaceStrings: Sendable, Equatable {
         reorderManyTabManagerUnavailable: String
     ) {
         self.closeProtected = closeProtected
+        self.closeFailed = closeFailed
         self.reorderManyMissingOrder = reorderManyMissingOrder
         self.reorderManyDuplicateWorkspace = reorderManyDuplicateWorkspace
         self.reorderManyWorkspaceNotFound = reorderManyWorkspaceNotFound
